gdataset/tests: don't spawn separate subprocess for test_datalist_clear() test
Yes, *if* the g_datalist_clear() is broken, then the test might deadlock
and hang.

However, it is not broken. The test will pass just fine, without spawning
a separate subprocess.

If we would spawn the test in a separate process every time, where a
test failure(!) might let it hang, we would almost always have to do
that.

Note that if the test really hangs, meson will just kill it after
timeout. That is a perfectly sensible behavior in case of the test
failure. So the failure will be noticed. Beyond that, we don't need to
program for the case where the test fails.

Also, if the test really would fail, it would be nice to just attach a
debugger and see where it's hanging. Spawning another process makes that
harder.

This was there since the beginning in commit d4be2925e43a ('Add a
testcase for the previous fix').